Share logic of filtered events between map and event list on org page.
