+++ src/utils/testing/mocks/mockState.ts
@@ -104,6 +104,11 @@ export default function mockState(overrides?: RootState) {
     },
     organizations: {
       eventsByOrgId: {},
+      filters: {
+        customDatesToFilterBy: [null, null],
+        dateFilterState: null,
+        orgIdsToFilterBy: [],
+      },
       orgData: remoteItem(0),
       subOrgsByOrgId: {},
       treeDataList: remoteList(),
